Push Segment #1
Revamped Menu minus the buttons due to an issue resizing the effects as they have text on them.

Fixed Resolution Scaling on Menu via the Menu Buttons.

Removed Audio Slider.

Several More Pushes In Bound. Was told to try pushing smaller increments by the Git Team in the mean time. If this one fails I will be setting up a new account and trying that.