Credits 'n' Changes
Added a Credits screen with a scrolling image of our names in a basic form to show at least what we are aiming for. Just needs a complete list put into an image and replaced to complete! (press A to skip back to main menu as well)

Fixed the Collectable buttons based on our new layout.

Next Push: Controller Screen and Language Selection Screen (unfunctional).